SQLExecutor for DuckDBTable into federation.rs
